requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.115.14
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
influxdb-client==1.44.0

# Environment variables
python-dotenv==1.0.1

# Security and password hashing
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# Additional dependencies for architecture
pydantic-settings==2.3.4

# SSH Tunnel management
psutil==6.1.0

# RabbitMQ client
pika==1.3.2
